+++ app/src/js/actions/index.js
@@ -203,6 +203,15 @@ export const listInactiveRequests = (options) => ({
   }
 });
 
+export const setWorkflowStep = (payload) => ({
+  [CALL_API]: {
+    type: types.REQUESTS,
+    method: 'POST',
+    path: 'data/submission/operation/changeStep',
+    body: payload
+  }
+});
+
 export const reviewRequest = (id, approve) => {
   return (dispatch) => {
     dispatch({
@@ -338,12 +347,12 @@ export const restoreRequest = (requestId) => ({
   }
 });
 
-export const addUserToRequest = (requestId, userId) => ({
+export const addUserToRequest = (payload) => ({
   [CALL_API]: {
-    type: types.REQUEST_RESTORE,
+    type: types.REQUEST_ADDUSER,
     method: 'POST',
-    path: 'data/submission/operation/restore',
-    body: { id: requestId }
+    path: 'data/submission/operation/addContributors',
+    json: payload
   }
 });
 
@@ -547,10 +556,9 @@ export const getOptionsUserGroup = () => ({
 export const getUser = (userId) => ({
   [CALL_API]: {
     type: types.USER,
-    id: userId,
     method: 'GET',
-    path: 'user/find',
-    qs: { id: userId }
+    id: userId,
+    path: `data/user/${userId}`
   }
 });